requirements.txt
appium-python-client==3.1.0
selenium==4.15.0
pytest==7.4.0
pytest-html==4.1.1